requirements/base.txt
rioxarray==0.14.0
dask==2023.1.1
xarray-spatial==0.3.5
pandera==0.14.5
shapely==2.0.1

###################################
# Potential future packages
#tqdm
#argcomplete
#scipy
#statsmodels
#geocube
#boto3
#pint-xarray (conversion of units)
#typer[all] - for CLI. alternative to argparse, Doesn't support Union types :/
# graphviz OR ipycytoscape - for dask graph visualizations
